Extract parser control-flow grammar
